+++ src/main/java/com/storm/score/controller/UserController.java
@@ -401,36 +401,4 @@ public class UserController {
     public ResponseEntity<com.storm.score.domain.user.User> signUp(@RequestBody com.storm.score.domain.user.User user) {
         return new ResponseEntity<>(user, HttpStatus.OK);
     }
-
-    @PostMapping("/login")
-    @ApiImplicitParams({
-            @ApiImplicitParam(name = "userId", value = "회원 아이디", required = true, dataType = "Long")
-    })
-    @ApiResponses(value = {
-            @ApiResponse(
-                    code = 200,
-                    message = "Successfully signed up",
-                    response = com.storm.score.domain.user.User.class,
-                    examples = @Example(
-                            @ExampleProperty(
-                                    mediaType = "application/json",
-                                    value = "- userId: 1\n  nickname: 경태\n  email: kt123@example.com\n"
-                            )
-                    )
-            )
-    })
-    public ResponseEntity<com.storm.score.domain.user.User> login(@RequestBody com.storm.score.domain.user.User user) {
-        return new ResponseEntity<>(user, HttpStatus.OK);
-    }
-
-    @PostMapping("/logout")
-    @ApiResponses(value = {
-            @ApiResponse(
-                    code = 204,
-                    message = "No Content"
-            )
-    })
-    public ResponseEntity<Void> logout() {
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-    }
 }
